Added a code to store all logs globally
All logs whether shown or not, should be kept. I’ve added a global
variable SkylinkJSLogs that will keep all of the log messages.

My reason for having his is so that I can better debug remote sessions.
It is better for me to upload the log to the server and view it there
than to ask for a screen share which may not be possible or reasonable.