fix(messaging): messages will not be sent under certain conditions
Messages that the user intended to be persisted was being sent as
encrypted messages (but not persisted) as a fallback for not meeting the persistent message
conditions, which may be misleading. SDK will throw error and drop the message in this scenario.

Only public messages can be persisted.

Fix block loop error not referencing the correct index in the listOfPeers array.

ESS-1902